stopped (most?) crashes

+++ ufo/src/kmint/force_driven_entities/population.cpp
@@ -2,10 +2,10 @@
 
 Population::Population(std::vector<kmint::ufo::human*>& h) : humans(h) {}
 
-void Population::saveProperties() { 
+void Population::saveProperties() {
     for (int i = 0; i < humans.size(); i++) {
         humanproperties[i].Alignment = humans[i]->AlignmentWeight;
-        humanproperties[i].Cohesion =  humans[i]->CohesionWeight;
+        humanproperties[i].Cohesion = humans[i]->CohesionWeight;
         humanproperties[i].Door = humans[i]->DoorWeight;
         humanproperties[i].Fitness = -1;
         humanproperties[i].GreenTank = humans[i]->GreenTankWeight;
@@ -13,12 +13,10 @@ void Population::saveProperties() {
         humanproperties[i].Separation = humans[i]->SeparationWeight;
         humanproperties[i].Ufo = humans[i]->UfoWeight;
     }
-
-
- }
+}
 
 void Population::Update() {
-    for (int i = 0; i < humans.size(); i++) {
+    for (unsigned long i = 0; i < humans.size(); i++) {
         if (humans[i]->removed()) {
             humans.erase(humans.begin() + i);
         }
@@ -30,7 +28,8 @@ void Population::getParents() {
     float totalFitness = 0;
     for (int i = 0; i < 100; i++) {
         totalFitness += humanproperties[i].Fitness;
-        std::cout << " Fitness of Human " << i << " is " << humanproperties[i].Fitness << std::endl;
+        std::cout << " Fitness of Human " << i << " is "
+                  << humanproperties[i].Fitness << std::endl;
     }
     std::cout << std::endl;
 
@@ -49,24 +48,24 @@ void Population::getParents() {
         parents[i] = std::make_pair(parent1_properties, parent2_properties);
     }
 
-        std::cout << std::endl;
+    std::cout << std::endl;
 }
 
 void Population::Crossover() {
     for (int i = 0; i < 100; i++) {
         HumanProperties p1 = parents[i].first;
         HumanProperties p2 = parents[i].second;
-        humans[i]->AlignmentWeight      = p1.Alignment;
-        humans[i]->CohesionWeight       = p1.Cohesion;
-        humans[i]->DoorWeight           = p1.Door;
-        humans[i]->GreenTankWeight      = p1.GreenTank;
-        humans[i]->RedTankWeight        = p2.RedTank;
-        humans[i]->SeparationWeight     = p2.Separation;
-        humans[i]->UfoWeight            = p2.Ufo;
+        humans[i]->AlignmentWeight = p1.Alignment;
+        humans[i]->CohesionWeight = p1.Cohesion;
+        humans[i]->DoorWeight = p1.Door;
+        humans[i]->GreenTankWeight = p1.GreenTank;
+        humans[i]->RedTankWeight = p2.RedTank;
+        humans[i]->SeparationWeight = p2.Separation;
+        humans[i]->UfoWeight = p2.Ufo;
     }
 }
 
-void Population::Mutation() { 
+void Population::Mutation() {
     std::cout << " Mutation" << std::endl;
     int MutationPercentage = 5;
 
@@ -76,20 +75,20 @@ void Population::Mutation() {
     std::cout << std::endl;
 }
 
-void Population::SetRandomMutation(int i){
+void Population::SetRandomMutation(int i) {
     int GenToMutate = RandomNumber(0, 7);
     std::cout << "Human " << i;
 
-    switch (GenToMutate) { 
+    switch (GenToMutate) {
         case 0:
             std::cout << " Cohesion mutated from " << humans[i]->CohesionWeight
                       << " to ";
-            humans[i]->CohesionWeight   = RandomNumber(0, 1);
+            humans[i]->CohesionWeight = RandomNumber(0, 1);
             std::cout << humans[i]->CohesionWeight << std::endl;
             break;
         case 1:
-            std::cout << " Separation mutated from " << humans[i]->SeparationWeight
-                      << " to ";
+            std::cout << " Separation mutated from "
+                      << humans[i]->SeparationWeight << " to ";
             humans[i]->SeparationWeight = RandomNumber(0, 1);
             std::cout << humans[i]->SeparationWeight << std::endl;
 
@@ -97,41 +96,41 @@ void Population::SetRandomMutation(int i){
         case 2:
             std::cout << " Alignment mutated from "
                       << humans[i]->AlignmentWeight << " to ";
-            humans[i]->AlignmentWeight  = RandomNumber(0, 1);
+            humans[i]->AlignmentWeight = RandomNumber(0, 1);
             std::cout << humans[i]->AlignmentWeight << std::endl;
 
             break;
         case 3:
-            std::cout << " RedTank mutated from "
-                      << humans[i]->RedTankWeight << " to ";
-            humans[i]->RedTankWeight    = RandomNumber(-1, 1);
+            std::cout << " RedTank mutated from " << humans[i]->RedTankWeight
+                      << " to ";
+            humans[i]->RedTankWeight = RandomNumber(-1, 1);
             std::cout << humans[i]->RedTankWeight << std::endl;
 
             break;
         case 4:
-            std::cout << " GreenTank mutated from " << humans[i]->GreenTankWeight
-                      << " to ";
-            humans[i]->GreenTankWeight  = RandomNumber(-1, 1);
+            std::cout << " GreenTank mutated from "
+                      << humans[i]->GreenTankWeight << " to ";
+            humans[i]->GreenTankWeight = RandomNumber(-1, 1);
             std::cout << humans[i]->GreenTankWeight << std::endl;
 
             break;
         case 5:
-            std::cout << " Ufo mutated from "
-                      << humans[i]->UfoWeight << " to ";
-            humans[i]->UfoWeight        = RandomNumber(-1, 1);
+            std::cout << " Ufo mutated from " << humans[i]->UfoWeight << " to ";
+            humans[i]->UfoWeight = RandomNumber(-1, 1);
             std::cout << humans[i]->UfoWeight << std::endl;
 
             break;
         case 6:
-            std::cout << " Door mutated from " << humans[i]->UfoWeight << " to ";
-            humans[i]->DoorWeight       = RandomNumber(-1, 1);
+            std::cout << " Door mutated from " << humans[i]->UfoWeight
+                      << " to ";
+            humans[i]->DoorWeight = RandomNumber(-1, 1);
             std::cout << humans[i]->DoorWeight << std::endl;
 
             break;
     }
 }
 
-std::pair<HumanProperties,int> Population::getParentsFromFitness(float total){
+std::pair<HumanProperties, int> Population::getParentsFromFitness(float total) {
     float p1 = RandomNumber(0, total);
 
     for (int i = 0; i < 100; i++) {
@@ -143,7 +142,6 @@ std::pair<HumanProperties,int> Population::getParentsFromFitness(float total){
 
     return std::make_pair(HumanProperties{0, 0, 0, 0, 0, 0, 0, 0}, 0);
 }
-
 
 void Population::increaseCohesion() {
     if (CohesionWeight <= 0.9f)
